requirements.txt
requests
beautifulsoup4
python-dateutil
feedgen
pandas
openpyxl
feedparser
reportlab
python-dotenv
readability-lxml
lxml